Code styling improved according to hints

+++ AdobeSdkService/HtmlToPdfConverter.cs
@@ -11,10 +11,11 @@ using Adobe.DocumentCloud.Services.pdfops;
 using Adobe.DocumentCloud.Services.io;
 using Adobe.DocumentCloud.Services.exception;
 using Adobe.DocumentCloud.Services.options.createpdf;
+using System.Threading.Tasks;
 
 namespace AdobeSdkService
 {
-    class HtmlToPdfConverter
+    internal class HtmlToPdfConverter
     {
         private readonly ILog log = LogManager.GetLogger(typeof(HtmlToPdfConverter));
         private readonly string inputFileNameOrUrl;
@@ -34,14 +35,12 @@ namespace AdobeSdkService
         /// Performs HTTP-request.
         /// </summary>
         /// <param name="url"></param>
-        /// <returns>Content stream.</returns>
-        private Stream MakeRequest(string url)
+        /// <returns>Response object.</returns>
+        private WebResponse MakeRequest(string url)
         {
             //  Defines code page and convert it to UTF-8.
-            var req = WebRequest.Create(url);
-            var res = req.GetResponse();
-
-            return res.GetResponseStream();
+            WebRequest req = WebRequest.Create(url);
+            return req.GetResponse();
         }
 
         /// <summary>
@@ -52,17 +51,17 @@ namespace AdobeSdkService
         private string CreateTemporaryZipFile(Stream content)
         {
             string tempPath = Path.GetTempPath();
-            string tempDirectoryName = tempPath + "/pdf_creator_page";
-            Directory.CreateDirectory(tempDirectoryName);
+            string tempDirectoryName = Path.Combine(tempPath, ReplaceInvalidCharacters(Path.GetRandomFileName()));
+            _ = Directory.CreateDirectory(tempDirectoryName);
 
             //  File must be named as "index.html".
             string tempFileName = Path.Combine(tempDirectoryName, "index.html");
-            using (var fileStream = File.Create(tempFileName))
+            using (FileStream fileStream = File.Create(tempFileName))
             {
                 content.CopyTo(fileStream);
             }
 
-            string zipFileName = Path.Combine(tempPath, Path.GetRandomFileName() + ".zip");
+            string zipFileName = Path.Combine(tempPath, ReplaceInvalidCharacters(Path.GetRandomFileName()) + ".zip");
             ZipFile.CreateFromDirectory(tempDirectoryName, zipFileName);
 
             File.Delete(tempFileName);
@@ -76,7 +75,7 @@ namespace AdobeSdkService
         /// </summary>
         public void CleanUp()
         {
-            if (urlIsProcessed)
+            if (urlIsProcessed && temporaryZipFileName != null)
             {
                 File.Delete(temporaryZipFileName);
             }
@@ -90,8 +89,10 @@ namespace AdobeSdkService
         {
             if (urlIsProcessed)
             {
-                var content = MakeRequest(inputFileNameOrUrl);
-                temporaryZipFileName = CreateTemporaryZipFile(content);
+                using (WebResponse response = MakeRequest(inputFileNameOrUrl))
+                {
+                    temporaryZipFileName = CreateTemporaryZipFile(response.GetResponseStream());
+                }
                 return FileRef.CreateFromLocalFile(temporaryZipFileName);
             }
             else
@@ -103,7 +104,7 @@ namespace AdobeSdkService
         /// <summary>
         /// Converts zip-file or page located by provided URL to PDF.
         /// </summary>
-        public void ConvertFileToPdf()
+        public async Task ConvertFileToPdfAsync()
         {
             try
             {
@@ -180,6 +181,23 @@ namespace AdobeSdkService
         {
             ILoggerRepository logRepository = LogManager.GetRepository(Assembly.GetEntryAssembly());
             XmlConfigurator.Configure(logRepository, new FileInfo("log4net.config"));
+        }
+
+        /// <summary>
+        /// Replaced inadmissible characters. Should necessarily replace '.' for proper naming of ZIP-files.
+        /// </summary>
+        /// <param name="fileName">Original file name.</param>
+        /// <returns>Transformed file name.</returns>
+        public static string ReplaceInvalidCharacters(string fileName)
+        {
+            char[] replacedCharacters = { '<', '>', ':', '"', '/', '\\', '|', '?', '*', '&', '#', '=', '.', '-' };
+
+            foreach (char character in replacedCharacters)
+            {
+                fileName = fileName.Replace(character, '_');
+            }
+
+            return fileName.Trim('_');
         }
     }
 }
